refactor: user wn slate avatar header in chat list screen
